🚚 Renamed `/global` to `/common`
