+++ src/components/common/NavBar.js
