+++ src/components/common/meta/getAuthorProperties.js
