+++ src/components/home/HomeHeader.js
@@ -2,9 +2,9 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import { Link } from 'gatsby'
 
-import NavBar from '../global/NavBar'
-import { SearchModal } from '../global/search'
-import { Icon } from '../global'
+import NavBar from '../common/NavBar'
+import { SearchModal } from '../common/search'
+import { Icon } from '../common'
 import { Spirit } from '../../styles/spirit-styles'
 
 class MainBox extends React.Component {
