+++ src/components/integrations/IntegrationBox.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import { Box } from '../global'
+import { Box } from '../common'
 
 const IntegrationBox = (props) => {
     const { post, hit } = props
