+++ src/components/integrations/IntegrationsContent.js
@@ -2,8 +2,8 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import _ from 'lodash'
 
-import { Layout } from '../global/layout'
-import NavBar from '../global/NavBar'
+import { Layout } from '../common/layout'
+import NavBar from '../common/NavBar'
 import IntegrationBox from "./IntegrationBox"
 import { Spirit } from '../../styles/spirit-styles'
 import { IntegrationIndex, IntegrationSearchBox, IntegrationResults } from './integration-search'
