+++ src/components/integrations/integration-search/IntegrationSearchBox.js
@@ -2,7 +2,7 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import { connectSearchBox } from 'react-instantsearch-dom'
 
-import { Icon } from '../../global'
+import { Icon } from '../../common'
 
 class IntegrationSearch extends React.Component {
     constructor(props) {
