+++ src/components/tutorials/PostCard.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types'
 
 import { Spirit } from '../../styles/spirit-styles'
 import getPostExcerpt from '../../utils/post-excerpt'
-import { Box } from '../global'
+import { Box } from '../common'
 
 const PostCard = ({ post, className }) => {
     // const tag = post.primaryTag ? post.primaryTag.name : (post.tags ? post.tags[0].name : `Untagged`)
