+++ src/pages/faq.js
@@ -2,10 +2,10 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import { graphql } from 'gatsby'
 
-import { Layout } from '../components/global/layout'
+import { Layout } from '../components/common/layout'
 import Link from 'gatsby-link'
 import { Spirit } from '../styles/spirit-styles'
-import { MetaData, getMetaImageUrls } from '../components/global/meta'
+import { MetaData, getMetaImageUrls } from '../components/common/meta'
 import { FAQTagList } from '../components/faq'
 
 class FAQPage extends React.Component {
